AzureDataLake.py
from azure.identity import DefaultAzureCredential
from azure.storage.blob import BlobServiceClient, BlobClient, ContainerClient

account_url = "https://victormrlucasformula1dl.blob.core.windows.net"
default_credential = DefaultAzureCredential()

# Create the BlobServiceClient object
blob_service_client = BlobServiceClient(account_url, credential=default_credential)
container_client = blob_service_client.get_container_client("hupe-docs")
blob_list = container_client.list_blobs()
for blob in blob_list:
    print("\t" + blob.name)
