requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#
python-dotenv
gunicorn
flask
flask-login
flask_sqlalchemy
pandas
pymongo[srv]
oauth2client
azure-identity
azure-keyvault-secrets
aiohttp
azure-storage-blob 